+++ src/text_to_image/domain/image_mode.py
