--- src/text_to_image/domain/text_image_saver.py
@@ -1,16 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import BinaryIO
-
-from .text_image import TextImage
-
-
-class TextImageSaver(ABC):
-    format: str
-
-    def __init__(self, format: str) -> None:
-        self.format = format
-
-    @abstractmethod
-    def save(self, fd: BinaryIO, image: TextImage) -> None:
-        """Save image to specified file.
-        Inherited classes must save image in own format"""
