--- src/text_to_image/pil_text_image_loader/__init__.py
@@ -1,4 +1,0 @@
-from .create_text_image_loader import create_text_image_loader
-from .pil_text_image_loader import PilTextImageLoader
-
-__all__ = ("PilTextImageLoader", 'create_text_image_loader')
