--- src/text_to_image/pil_text_image_saver/__init__.py
@@ -1,4 +1,0 @@
-from .create_text_image_saver import create_text_image_saver
-from .pil_text_image_saver import PilTextImageSaver
-
-__all__ = ('create_text_image_saver', 'PilTextImageSaver')
